Fixes for the new bulk execution backend (#31884)
